pyproject.toml
[project]
name = "kakochessbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chess>=1.11.2",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "numpy>=1.26.0",
    "torch>=2.11.0",
    "uvicorn>=0.44.0",
    "onnxscript>=0.6.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
